redesign app layout: removed navbar and added profile menu to sidebar
