--- Client/src/Components/NavBar/index.css
@@ -1,34 +1,0 @@
-/* NavBar Component Styles*/
-.MuiToolbar-root {
-  min-height: var(--env-var-nav-bar-height) !important;
-}
-
-.icon-button-toggle-title {
-  font-size: var(--env-var-font-size-medium);
-  color: var(--env-var-color-5);
-}
-
-.icon-button-toggle-pic {
-  width: 10px;
-  height: 5px;
-}
-
-#icon-button {
-  -webkit-transition: none;
-  transition: none;
-  outline: none;
-  box-shadow: none;
-  background-color: transparent;
-}
-
-#menu-appbar svg {
-  width: 16px;
-  height: 16px;
-  color: var(--env-var-color-25);
-}
-
-#bw-uptime-logo-dashboard {
-  width: fit-content;
-  height: 16px;
-  margin-left: 15px;
-}
