+++ Client/src/Components/Sidebar/index.css
@@ -18,11 +18,14 @@ aside span.MuiTypography-root {
   color: var(--env-var-color-5);
   line-height: 1;
 }
-aside .MuiStack-root:nth-last-child(2) {
+aside .MuiStack-root:nth-last-child(3) {
   margin-top: auto;
+}
+aside .MuiStack-root:last-child,
+aside .MuiStack-root:nth-last-child(3) {
   position: relative;
 }
-aside .MuiStack-root:nth-last-child(2):before {
+aside .MuiStack-root:last-child:before {
   content: "";
   position: absolute;
   top: -10px;
@@ -31,5 +34,35 @@ aside .MuiStack-root:nth-last-child(2):before {
   border-top: solid 1px var(--env-var-color-6);
 }
 aside .MuiStack-root:last-child {
-  margin-bottom: 10px;
+  margin-top: 15px;
+}
+
+.sidebar-menu {
+  margin-top: -20px;
+}
+.sidebar-menu .MuiPaper-root {
+  box-shadow: var(--env-var-shadow-1);
+  border: solid 1px var(--env-var-color-6);
+  border-radius: var(--env-var-radius-1);
+  gap: 1px;
+}
+.sidebar-menu .MuiList-root {
+  min-width: 100px;
+  width: 150px;
+  padding-bottom: 0;
+}
+.sidebar-menu li.MuiButtonBase-root:last-child {
+  border-top: solid 1px var(--env-var-color-6);
+  padding: 12px 16px;
+}
+.sidebar-menu li.MuiButtonBase-root {
+  min-height: fit-content;
+}
+.sidebar-menu .MuiList-root svg {
+  width: 16px;
+  height: 16px;
+}
+.sidebar-menu span {
+  font-size: var(--env-var-font-size-medium);
+  color: var(--env-var-color-2);
 }
